LayoutTests/platform/wpe/storage/domstorage/localstorage/simple-usage-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
This test tries simple operations on localStorage
Length is 0
Value for FOO is null
